requirements.txt
# Telegram Bot Dependencies
python-telegram-bot==20.7
telegram==0.0.1

# Google AI Studio / Gemini API
google-generativeai==0.3.2

# Environment and Configuration
python-dotenv==1.0.0

# HTTP Client
aiohttp==3.9.1
requests==2.31.0

# Logging and Utilities
colorlog==6.8.0

# Development Dependencies (optional)
pytest==7.4.3
pytest-asyncio==0.23.2
black==23.12.0
flake8==6.1.0
